fixup! [FLINK-10309][rest] Before shutting down cluster, wait for asynchronous operations.
